Add new configurable LocalDaskExecutor and change default to LocalExecutor
