+++ tests/engine/test_defaults.py
@@ -4,21 +4,21 @@ from prefect import engine, utilities
 
 
 def test_default_executor():
-    assert engine.get_default_executor_class() is engine.executors.SynchronousExecutor
+    assert engine.get_default_executor_class() is engine.executors.LocalExecutor
 
 
 def test_default_executor_responds_to_config():
     with utilities.configuration.set_temporary_config(
-        {"engine.executor.default_class": "prefect.engine.executors.LocalExecutor"}
+        {"engine.executor.default_class": "prefect.engine.executors.LocalDaskExecutor"}
     ):
-        assert engine.get_default_executor_class() is engine.executors.LocalExecutor
+        assert engine.get_default_executor_class() is engine.executors.LocalDaskExecutor
 
 
 def test_default_executor_responds_to_config_object():
     with utilities.configuration.set_temporary_config(
-        {"engine.executor.default_class": engine.executors.LocalExecutor}
+        {"engine.executor.default_class": engine.executors.LocalDaskExecutor}
     ):
-        assert engine.get_default_executor_class() is engine.executors.LocalExecutor
+        assert engine.get_default_executor_class() is engine.executors.LocalDaskExecutor
 
 
 def test_default_executor_with_bad_config():
@@ -26,10 +26,7 @@ def test_default_executor_with_bad_config():
         {"engine.executor.default_class": "prefect.engine.bad import path"}
     ):
         with pytest.warns(UserWarning):
-            assert (
-                engine.get_default_executor_class()
-                is engine.executors.SynchronousExecutor
-            )
+            assert engine.get_default_executor_class() is engine.executors.LocalExecutor
 
 
 def test_default_flow_runner():
